move cert/key fixture to conftest
